refactor: 重构类型定义
