+++ src/components/page-draggable/index.vue
@@ -3,7 +3,7 @@
     <kzy-draggable
       v-bind="props.draggableConfig.draggableProp"
       :data-list="dataList"
-      @handle-change="props.draggableConfig?.handleChangeF"
+      @handle-change="props.draggableConfig?.handleChange"
     >
       <template #item="{ element }">
         <div class="page-drag-slot">
@@ -16,7 +16,7 @@
 
 <script lang="ts" setup>
 import type { Draggable } from '@/components/base-ui/kzy-draggable/types'
-import type { LibraryComponent } from '@/library/types'
+import type { LibraryComponent } from '@/types/library-component'
 import KzyDraggable from '@/components/base-ui/kzy-draggable/index.vue'
 
 const props = defineProps<{
