+++ src/library/swipe/index.vue
@@ -16,17 +16,19 @@
 <script lang="tsx">
 import { ElIcon } from 'element-plus'
 import { Image as VanImage } from 'vant'
-import { LibraryNameEnum } from '@/views/library-panel/types'
+import {
+  AttributePanelFormItemInputTypeEnum,
+  AttributePanelsEnum,
+  LibraryPanelTabEnum,
+} from '@/types/panel'
 import { createLibraryComponentPropItem, defineLibraryComponent } from '@/utils/library'
-import { AttributePanelsEnum } from '@/views/attribute-panel/types'
-import { EditableConfigItemInputTypeEnum } from '@/views/edit-panel/types'
 import preview from '@/library/swipe/components/preview.vue'
-import { LibraryComponentFormItemLabelPositionEnum } from '@/library/types'
+import { LibraryComponentFormItemLabelPositionEnum } from '@/types/library-component'
 
 export default defineComponent({
   ...defineLibraryComponent({
     name: 'LibSwipe',
-    libraryName: LibraryNameEnum.generics,
+    libraryName: LibraryPanelTabEnum.generics,
     tabName: 'form',
     order: 3,
     libraryPanelShowDetail: {
@@ -48,7 +50,7 @@ export default defineComponent({
   props: {
     urlList: createLibraryComponentPropItem({
       title: '图片链接列表',
-      formType: EditableConfigItemInputTypeEnum.indefiniteNumberInputBox,
+      formType: AttributePanelFormItemInputTypeEnum.indefiniteNumberInputBox,
       belongToPanel: AttributePanelsEnum.generic,
       type: Array,
       default: [
@@ -62,56 +64,56 @@ export default defineComponent({
     }),
     initialSwipe: createLibraryComponentPropItem({
       title: '初始位置索引值',
-      formType: EditableConfigItemInputTypeEnum.input,
+      formType: AttributePanelFormItemInputTypeEnum.input,
       belongToPanel: AttributePanelsEnum.generic,
       default: '0',
       type: [String, Number],
     }),
     autoplay: createLibraryComponentPropItem({
       title: '自动轮播间隔',
-      formType: EditableConfigItemInputTypeEnum.input,
+      formType: AttributePanelFormItemInputTypeEnum.input,
       belongToPanel: AttributePanelsEnum.generic,
       default: '3000',
       type: [String, Number],
     }),
     duration: createLibraryComponentPropItem({
       title: '动画时长',
-      formType: EditableConfigItemInputTypeEnum.input,
+      formType: AttributePanelFormItemInputTypeEnum.input,
       belongToPanel: AttributePanelsEnum.generic,
       default: '500',
       type: [String, Number],
     }),
     loop: createLibraryComponentPropItem({
       title: '循环播放',
-      formType: EditableConfigItemInputTypeEnum.switch,
+      formType: AttributePanelFormItemInputTypeEnum.switch,
       belongToPanel: AttributePanelsEnum.generic,
       default: true,
       type: Boolean,
     }),
     showIndicators: createLibraryComponentPropItem({
       title: '显示指示器',
-      formType: EditableConfigItemInputTypeEnum.switch,
+      formType: AttributePanelFormItemInputTypeEnum.switch,
       belongToPanel: AttributePanelsEnum.generic,
       default: true,
       type: Boolean,
     }),
     picWidth: createLibraryComponentPropItem({
       title: '图片宽度',
-      formType: EditableConfigItemInputTypeEnum.input,
+      formType: AttributePanelFormItemInputTypeEnum.input,
       belongToPanel: AttributePanelsEnum.generic,
       default: '100%',
       type: [String, Number],
     }),
     picHeight: createLibraryComponentPropItem({
       title: '图片高度',
-      formType: EditableConfigItemInputTypeEnum.input,
+      formType: AttributePanelFormItemInputTypeEnum.input,
       belongToPanel: AttributePanelsEnum.generic,
       default: '240',
       type: [String, Number],
     }),
     lazyRender: createLibraryComponentPropItem({
       title: '延迟渲染未展示的轮播',
-      formType: EditableConfigItemInputTypeEnum.switch,
+      formType: AttributePanelFormItemInputTypeEnum.switch,
       belongToPanel: AttributePanelsEnum.generic,
       default: false,
       type: Boolean,
